+++ lib/src/actions/base_actions.dart
@@ -4,3 +4,4 @@ part 'actions/checks.dart';
 part 'actions/explosion.dart';
 part 'actions/hands.dart';
 part 'actions/immortality.dart';
+part 'actions/points.dart';